==> 1_MonitoringArea.Rmd <==
---
title: "About the Monitoring Area"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE,
                      fig.height = 5, fig.width = 8
                      )  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
      detach, character.only=TRUE, unload=TRUE, force=TRUE))
}

pacman::p_load(knitr, kableExtra, 
               ggpubr, tidyverse,
               # for kaya - mapping
               ggmap, sf, #mapping
               ggspatial, #mapping, adding scales, N arrows...
               spData, # us_states data - mapping WA state
               cowplot, #layering ggplots. ggdraw(), draw_plot()
               units #convert between e.g., m to km
               )    
 

set.seed(1)

theme_set(theme_bw())
theme_update(legend.position = "bottom")

```

# upload data 

 

```{r}
gis_path0 <- file.path("~", "OneDrive - UW", "Documents", "School", "PhD", "Dissertation", "GIS", "Shapefiles") 
gis_path <- file.path("..",  "OneDrive - UW", "GIS", "Shapefiles") 
data_path <- file.path("~", "OneDrive - UW", "Documents", "School", "PhD", "Dissertation", "TRAP R Project", "Data")

project_crs <- 4326  #lat/long
m_crs <- 32148 #meters

```

```{r}
#point locations

act0 <- readRDS(file = file.path(data_path, "Aim 2", "Geocovariates", "cov_act_preprocessed.rda"))

#add stop & route number crosswalk
route_num <- read.csv(file.path(data_path, "Aim 2", "Mobile Monitoring", "locations_190715.csv")) %>%
  mutate(Route = substr(route,str_length(route), str_length(route))) %>%
  select(native_id = site_id, Route)

mm <- readRDS(file.path(data_path, "Aim 2", "Geocovariates", "cov_mm_preprocessed.rda")) %>%
  # drop pop90, pop_ ( & ??NDVI)
  select(-contains(c("pop90_", "pop_"))) %>%
  #drop Roosevelt garage & stop w/ 1 obeservation that was replaced by MS0601
  filter(!site_id %in% c("MS0000", "MS0398")) %>%
  mutate(
    Kind = factor(ifelse(grepl("MS", native_id), "Monitoring Stop", "AQS Colocation"), 
                  levels = c("Monitoring Stop", "AQS Colocation") 
                  )
  ) %>%
  left_join(route_num)


drop_parameters <- c("Temperature", "Wind", "Pressure", "Relative Humidity", "Sample Flow", "Sample Volume")

aqs_all0 <- read.csv(file.path("Data", "EPA_annual_conc_by_monitor_2020.csv")) %>%
  filter(
    #only keep those in this study area
    grepl("Seattle", CBSA.Name),
    #sites that only have meteorological or sample paramters
    !grepl(paste0(drop_parameters, collapse = "|"), Parameter.Name, ignore.case = T)
    ) %>%
  mutate(
    #make new ID
    native_id = paste0(State.Code,
             str_pad(string =  County.Code,
                     width = 3, 
                     side = "left", 
                     pad = 0),
             str_pad(string =  Site.Num,
                     width = 4, 
                     side = "left", 
                     pad = 0)
             ),
  )


#all AQS sites in Seattle area
aqs_all <- aqs_all0  %>%
  distinct(native_id, #State.Code, County.Code, Site.Num, #Parameter.Name,
           Latitude, Longitude
           )

#aqs sites included in study
aqs <- readRDS(file = file.path(data_path, "Aim 2", "Geocovariates", "cov_agency_preprocessed.rda"))

```

```{r}

study_area_shp <- read_sf(file.path(gis_path0, "Study area", "oval_around_monitoring_area.shp")) %>%
  st_transform(project_crs)

routes_shp <- read_sf(file.path(gis_path0, "Routes 190910", "All routes.shp")) %>%
  st_transform(project_crs) %>%
  mutate(
    Route = as.factor(Route)
  )

# monitoring, study, area shapefiles
monitoring_area_shp <- read_sf(file.path(
  gis_path0, "Study Area", "monitoring_area_filled_in.shp"
  #gis_path, "Monitoring", "monitoring_area_filled_in.shp"
  )) %>%
  st_transform(project_crs)

## make "monitoring area" slightly larger so that all driving routes also fall inside
r7_north <- routes_shp %>%
  filter(Route == 7) %>%
  slice(3) %>%
  st_transform(m_crs) %>%
  st_buffer(1300) %>%
  st_transform(project_crs)

r6_south <- routes_shp %>%
  filter(Route == 6) %>%
  slice(4) %>%
  st_transform(m_crs) %>%
  st_buffer(1500) %>%
  st_transform(project_crs) #%>% st_union() 

monitoring_area_shp2 <- routes_shp %>%
  filter(Route %in% c(9, 8)) %>%
  st_transform(m_crs) %>%
  st_buffer(1500) %>%
  st_transform(project_crs) %>%
  # ? make buffer into a single featuer
  st_union() %>%
  # combine buffered routes with original monitoring area
  st_union(monitoring_area_shp) %>%
  st_union(r6_south) %>%
  st_union(r7_north)

saveRDS(monitoring_area_shp2, file.path("Data", "Output", "GIS", "monitoring_area_shp.rda"))

# water file with all bodies of water in the monitoring region (for predicting in_monitoring_region=TRUE)
water_shp <- read_sf(file.path("~", "OneDrive - UW", "Documents", "School", "PhD", "Dissertation", "GIS", "Shapefiles", "Other features", "Water", "DNR_Hydrography__Water_Bodies", "DNR_Hydrography__Water_Bodies.shp")) %>%
  st_transform(project_crs) %>%
  # make water feature valid - otherwise has plotting/other issues
  st_make_valid() %>% #st_buffer(0)
  # only keep water in the region - make file smaller
  st_intersection(., st_transform(st_buffer(st_transform(monitoring_area_shp2, m_crs), 10e3), project_crs)) 

saveRDS(water_shp, file.path("Data", "Output", "GIS", "all_water_nearby_shp.rda"))

# PREDICTION shapefile: monitoring land without any bodies of water
monitoring_land_zero_water_shp <- st_difference(monitoring_area_shp2, st_union(water_shp))
saveRDS(monitoring_land_zero_water_shp, file.path("Data", "Output", "GIS", "monitoring_land_zero_water_shp.rda"))
#plot(monitoring_land_zero_water_shp)

# water file w/ only large bodies of water (for mapping/visualizations)
water_shp1 <- water_shp %>% 
  # only keep large bodies of water
  filter(grepl("Puget Sound|Lake Washington|Lake Sammamish|Lake Union", WB_GNIS_NM)) %>%
  st_union()

saveRDS(water_shp1, file.path("Data", "Output", "GIS", "water_nearby_shp.rda"))

# MAPPING shapefile: monitoring land without large bodies of water
monitoring_land <- st_difference(monitoring_area_shp2, st_union(water_shp1))  
saveRDS(monitoring_land, file.path("Data", "Output", "GIS", "monitoring_land_shp.rda"))
#plot(monitoring_land)

```



```{r}
#plot(monitoring_area_shp$geometry)


# test
# 
# water0 <- read_sf(file.path("..", "GIS", "Shapefiles", "Other features", "Open_water_for_King_County_and_portions_of_adjacent_counties___wtrbdy_area", "Open_water_for_King_County_and_portions_of_adjacent_counties___wtrbdy_area.shp"))
# 
# # resulting file is larger than before
# water0.1 <- water0 %>% 
#   st_make_valid() %>% #st_buffer(0)
#   # only keep water in the region - make file smaller
#   st_intersection(., st_transform(st_buffer(st_transform(monitoring_area_shp2, m_crs), 10e3), project_crs))


```




size of monitoring Region

```{r}
"entire monitoring region, with water"
## 1373.504 [km^2] 
st_area(monitoring_area_shp2) %>%
  set_units("km^2")


"monitoring area without water" #: 1200.383 [km^2]
monitoring_land %>%
  st_area() %>%
  set_units("km^2")

monitoring_land %>%
  st_area() %>%
  set_units("mi^2")


# # check that things look correct. # looks good
# ggplot() +
#   #geom_sf(data=water_shp1, aes(fill="water"), alpha=0.3) +
#   #geom_sf(data=monitoring_area_shp2, aes(col="original"), alpha=0.15) +
#   geom_sf(data=monitoring_land, aes(fill="monitoring land"), alpha=0.7)  
   
```


```{r}

act <- act0 %>%
  #drop duplicate locations
  select(-native_id, -site_id) %>%
  distinct() %>%
  #add new IDs for unique locations
  mutate(native_id = paste0("act_",
                            str_pad(row_number(), width = 5, side = "left", pad=0)
                            )
         ) %>%
  select(native_id, everything())  

#create sf object for locations
act_shp <- act %>%
  #make sf object
  st_as_sf(., coords=c("longitude","latitude"), remove=F,
           crs=project_crs)

mm_shp <- mm %>%
  st_as_sf(., coords=c("longitude","latitude"), remove=F,
           crs=project_crs)

aqs_shp <- aqs %>%
  st_as_sf(., coords=c("longitude","latitude"), remove=F,
           crs=project_crs)

aqs_all_shp <- aqs_all %>%
  st_as_sf(., coords=c("Longitude","Latitude"), remove=F,
           crs=project_crs)



# cohort locations in monitoring and study areas
act_shp$in_monitoring_area <- st_intersects(act_shp, monitoring_area_shp, sparse = F) %>%
  apply(., 1, any)
act_shp$in_study_area <- st_intersects(act_shp, study_area_shp, sparse = F) %>%
  apply(., 1, any)

aqs_shp$in_monitoring_area <- st_intersects(aqs_shp, monitoring_area_shp, sparse = F) %>%
  apply(., 1, any)  

aqs_all_shp$in_monitoring_area <- st_intersects(aqs_all_shp, monitoring_area_shp, sparse = F) %>%
  apply(., 1, any)  


#clean/convert back to df

aqs_shp <- aqs_shp %>%
  filter(in_monitoring_area==TRUE) %>%
  select(-in_monitoring_area)

aqs <- aqs_shp %>%
  st_drop_geometry() #%>%
  # filter(in_monitoring_area==TRUE) %>%
  # select(-in_monitoring_area)

aqs_all_shp <- aqs_all_shp %>%
  filter(in_monitoring_area==TRUE) %>%
  select(-in_monitoring_area)

aqs_all <- aqs_all_shp %>%
  st_drop_geometry() #%>%
  # filter(in_monitoring_area==TRUE) %>%
  # select(-in_monitoring_area)


```

proportion of ACT locations in study and monitoring areas

```{r}

act_shp %>%
  st_drop_geometry() %>%
  group_by(in_monitoring_area) %>%
  summarize(
    n = n(),
    all_act_locations = nrow(.),
    prop = n/all_act_locations
  ) %>%
  kable(caption = "unique ACT locations in the monitoring area", 
        digits = 2) %>% 
  kable_styling()
  
```

```{r}
#only keep shapefile w/ sites inside monitoring area

act_shp <- act_shp %>%
  #only need these points
  filter(in_monitoring_area == TRUE) %>%
  select(-in_monitoring_area, -in_study_area) 
 
act <- act_shp %>%
  st_drop_geometry() #%>% 
  ##only need these points
  # filter(in_monitoring_area == TRUE) %>%
  # select(-in_monitoring_area, -in_study_area) 


```

```{r, eval=F}
#save files as shp files for easy QGIS mapping verification later

#save cohort locations 
act_shp %>%
  select(native_id, latitude, longitude) %>%
  st_write(file.path(gis_path, "ACT", paste0("unique_act_locations_", Sys.Date() , ".shp")) )

# save mobile monitoring stop locations
mm_shp %>%
  select(native_id, latitude, longitude) %>%
  #mutate(
  #   kind = ifelse(grepl("MS", native_id), "Stop", "AQS Colocation")
  # ) %>%
  st_write(file.path(gis_path, "Monitoring", paste0("mm_locations", ".shp")) )

#save all AQS sites
aqs_all_shp %>%
  select(native_id, Latitude, Longitude) %>%
  st_write(file.path(gis_path, "AQS Sites", paste0("all_aqs_sites", ".shp")), 
           #delete file if already exists
           delete_layer=TRUE )


```



# Jitter Cohort locations

add new columns for this

```{r}
#act_shp_jit

#doestn' plot anything other than the main coordinates if keep as same sf object?
act_shp_jit <- act_shp %>%
  #jitter in a meters transformation
  st_transform(m_crs) %>%
  mutate(
     long_m = st_coordinates(.)[,1],
     lat_m = st_coordinates(.)[,2],
     
     #jittered. max amount that locations should be jittered: +-200 m
     lat_jit_m = jitter(lat_m, amount = 500),
     long_jit_m = jitter(long_m, amount = 500)
     ) %>% 
  #convert new meter coordinates to lat/long
  st_drop_geometry() %>%
  st_as_sf(., coords=c("long_jit_m","lat_jit_m"), remove=F,
           crs=m_crs) %>%
  #get these in lat/long (like other shapefiles - for plotting)
  st_transform(project_crs) %>%
  mutate(
    long_jit_deg = st_coordinates(.)[,1],
    lat_jit_deg = st_coordinates(.)[,2],
  ) #%>%
  #back to normal
  # st_drop_geometry() %>%
  # st_as_sf(., coords=c("longitude","latitude"), remove=F,
  #          crs=project_crs)
   
  

  


```


# Distance between monitoring and cohort locations

AQS sites are all the ones in the monitoring area (n=9)


```{r}
#calculate distance from cohort locations to monitors
# example: https://dominicroye.github.io/en/2020/geographic-distance/

dist_aqs <- st_distance(act_shp, aqs_all_shp) %>%
  #set_units(km) %>%
   as.data.frame()

dist_mm <- st_distance(act_shp, mm_shp) %>%
  #set_units(km) %>%
   as.data.frame()

#find monitor w/ shortest distance
pos_aqs <- apply(dist_aqs, 1, which.min)
pos_mm <- apply(dist_mm, 1, which.min)
#calculate shortest distance (meters)
min_dist_aqs <- apply(dist_aqs, 1, min)
min_dist_mm <- apply(dist_mm, 1, min)


act_shp <- act_shp %>%
  mutate(
    #nearest_monitor_type = all_monitoring$monitoring[pos],
    nearest_aqs_monitor = aqs_all_shp$native_id[pos_aqs],
    nearest_aqs_lat = aqs_all_shp$latitude[pos_aqs],
    nearest_aqs_long = aqs_all_shp$longitude[pos_aqs],
    
    nearest_mm_monitor = mm_shp$native_id[pos_mm],
    nearest_mm_lat = mm_shp$latitude[pos_mm],
    nearest_mm_long = mm_shp$longitude[pos_mm],
    
    dist_aqs_monitor = min_dist_aqs,
    dist_mm_monitor = min_dist_mm

  )


act_shp %>%
  st_drop_geometry() %>% 
  gather("Monitor", "distance", dist_aqs_monitor, dist_mm_monitor) %>% 
  mutate(Monitor = ifelse(grepl("aqs", Monitor), "AQS", "Mobile")) %>%  
  
  group_by(Monitor) %>%
  summarize(
    ACT_N = n(),
    Min = min(distance),
    Q05 = quantile (distance, 0.05),
    Mean = mean(distance),
    Sd = sd(distance),
    Q95 = quantile (distance, 0.95),
    Max = max(distance)
  ) %>%
  kable(caption = "distance (m) from AQS (n=6) and mobile stop (n=309) monitoring locations to ACT cohort locations in the monitoring area", 
        digits = 0
        ) %>%
  kable_styling()
  
```


# Map


```{r}
# create background map for grid/mapping
#bbox <- st_bbox(act_shp)

#make box little bigger than monitoring area
bbox <- st_bbox(st_transform(st_buffer(st_transform(act_shp, m_crs), 10000), project_crs ))

names(bbox) <- c("left", "bottom", "right", "top")

# background map
map0 <- suppressMessages(get_stamenmap(bbox = bbox, 
                                      zoom = 11,
                                      #maptype = "toner" #"toner-lite"
                                      maptype = "terrain" 
                                      )
                         ) %>%
  # Make basic map image from the tiles
  ggmap(ggmap = ., darken = c(0.5, "white")) + theme_void()

## usage example: 
# map0 + geom_point()
```


 

```{r, fig.height=10}

map1 <- map0 + 
  #monitoring area
  geom_sf(data = monitoring_area_shp2, 
                aes(fill = "Monitoring Area"), 
          inherit.aes = F,

          #reduce/eliminate outline
          lwd = 0.1, #0, 
          alpha = 0.1,
           ) + 
  
  #jittered ACT locations
  geom_point(data = act_shp_jit, 
                aes(
                  shape="ACT Cohort",
                    y=lat_jit_deg,
                    x=long_jit_deg
                    ),
              
          inherit.aes = F,
          alpha=0.2,
          size=0.5,
          #shape=5
           ) +
  
  #route lines
  geom_sf(data=routes_shp,
        aes(col = Route
            ),
        inherit.aes = F,
        alpha=0.7
        ) +
  
  #monitoring stops
  geom_sf(data = mm_shp, 
          aes(col = Route,  
              ),
          inherit.aes = F,
          size=2  
           ) + 
  #colocations
  geom_sf(data = filter(mm_shp, Kind == "AQS Colocation"), 
          aes(shape= "AQS Colocation"),
          inherit.aes = F,
          size=3 
           ) + 
  
    # # add scale & N arrow to top
  geom_sf(data = monitoring_area_shp2, inherit.aes = FALSE,
          #don't actually the polygon our outline
          alpha=0, lwd = 0) +
  annotation_scale(data = monitoring_area_shp2, location = "tl") +
  annotation_scale(data = monitoring_area_shp2, location = "tl", unit_category ="imperial", pad_y = unit(0.55, "cm")) +
    annotation_north_arrow(location = "tl",
                           #point towards North Pole
                           which_north = "true",
                           pad_y = unit(0.5, "in"),
                           style = north_arrow_fancy_orienteering
                           ) +

  theme_bw() +
  
  theme(
      legend.justification=c(1,1), #(0,1)
      legend.position=c(1,1),#(0,1)
      legend.background =  element_blank()
      ) + 
  
  scale_shape_manual(values = c(2,1 #,21
                                )) +
  
  #order how legend items are presented
  guides(col = guide_legend(order = 2),
         shape = guide_legend(order = 3), 
         fill = guide_legend(order = 1)
         
         ) +
  
  #add attribution/reference to bottom left
  geom_text(aes(x=-Inf, y=-Inf, hjust=-0.01, vjust=-0.5,
                label= "Map tiles by Stamen Design, under CC BY 3.0. \nData by OpenStreetMap, under ODbL."
                ),
            size=3
            ) +
  
    labs( 
         x = "Longitude",
         y = "Latitude",
         fill = "",
         col = "Monitoring\nRoute",
         shape = "",  
         caption = ""
         ) 

#map1

```


add inset map 

```{r, fig.height=10}
# example code: https://geocompr.github.io/post/2019/ggplot2-inset-maps/ 

data("us_states", package = "spData")

wa_map <- us_states %>%
  filter(NAME == "Washington") %>%
  #make sure this is in the same crs 
  st_transform(project_crs)

wa_centroid <- st_coordinates(st_centroid(wa_map))

# #make bbox and sf object of a rectangular box 
# bbox0 <- st_as_sfc(bbox) 

inset_map <- ggplot() + 
  geom_sf(data=wa_map, fill = "white", alpha=0.5) + 
  #geom_sf(data=bbox0, fill = "white", col="blue", alpha=0.5) + 
  geom_sf(data = monitoring_area_shp2, #aes(), 
          fill = "pink",
          #reduce/eliminate outline
          lwd = 0.1,  
          alpha = 0.7,
           ) +
  theme_void() + 
  #add "WA" label
  geom_text(aes(x = wa_centroid[1], y = wa_centroid[2]),
            label = "WA", 
            size=4
                )
  
# inset_map

# use cowplot to layer both maps
ggdraw() +
  draw_plot(map1) +
  draw_plot(inset_map, 
            # The distance along a (0,1) x- or y- axis to draw the left/bottom edge of the plot
            x = 0.65, y = 0.005, 
            #x = 0.25, y = 0.005, 
            # The width and height of the plot expressed as proportion of the entire ggdraw object
            width = 0.23, height = 0.23)


ggsave(file.path("..", "Manuscript", "Images", "monitoring_map.jpg"), width = 7.5, height = 10)

```

## Abstract/TOC art

title: "Mobile monitoring of traffic-related air pollution for a prospective cohort study in the greater Seattle area"

* zoom in to primary map; make simpler (e.g., no legends); use [UW colors - purple dots](https://www.washington.edu/brand/graphic-elements/primary-color-palette/)      
* drop car; shorten text






```{r, eval=F }

## old option 

# create background map for grid/mapping
one_route_shp <- routes_shp %>%
  filter(Route %in% c(1,2, 9))

#make box little bigger than monitoring area
bbox2 <- st_bbox(st_transform(st_buffer(st_transform(one_route_shp, m_crs), 1000), project_crs ))

names(bbox2) <- c("left", "bottom", "right", "top")

# background map
map0_zoom <- suppressMessages(get_stamenmap(bbox = bbox2, 
                                      zoom = 10, 
                                      #maptype = "terrain-lines" #
                                     maptype = "toner-background" #background #lines
                                       
                                      )
                         ) %>%
  # Make basic map image from the tiles
  ggmap(ggmap = ., #darken = c(0.5, "white")
        #UW gold background
        darken = c(0.4,  "#b7a57a")
        ) + theme_void()

## usage example: 
#map0_zoom #%>% ggmap()  

 

map0_zoom + 
  #route lines
  geom_sf(data=routes_shp,inherit.aes = F,
          #UW purple 
          col = "#4b2e83", #alpha=0.7
        ) +
  
  #monitoring stops
  geom_sf(data = mm_shp, inherit.aes = F,
            col = "#4b2e83", size=2  
           )  

```

## option 1

```{r}

#basic outline shapefile for near UW study area
kc_shp <- read_sf(file.path(
  "..", "GIS", "Shapefiles", "Other features","Congressional_Districts_in_King_County___congdst_area", "Congressional_Districts_in_King_County___congdst_area.shp"
  )) %>%
  st_transform(project_crs) %>%
  #make into one polygon
  st_union()

#plot(kc_shp) # $geometry

```

```{r}
one_route_shp2 <- routes_shp %>%
  filter(Route %in% c(1, 2, 3, 9))

bbox3 <- st_bbox(one_route_shp2) %>% as.vector()

#modify xmin, xmax & ymax 
bbox3[1] <- st_bbox(one_route_shp2)[1]*1.0003  %>% as.vector()
bbox3[2] <- bbox3[2]*1.0013 %>% as.vector()  
bbox3[3] <- st_bbox(one_route_shp2)[3]*.9985 %>% as.vector()  
bbox3[4] <- st_bbox(kc_shp)[4]*.9995 %>% as.vector()

text_size <- 2.25#2.5

ggplot() +
  geom_sf(data=kc_shp, fill = "#b7a57a", alpha=0.3, size=0.1) +
  #route lines
  geom_sf(data=routes_shp,inherit.aes = F,
          #UW purple 
          col = "#4b2e83",  size=0.2, alpha=0.6#linetype=3 
        ) +
  #monitoring stops
  geom_sf(data = mm_shp, inherit.aes = F,
            col = "#4b2e83", size=0.5, alpha=0.8
           )  + 
  coord_sf(xlim = bbox3[c(1,3)], ylim = bbox3[c(2,4)], expand = FALSE) +
  
  geom_label(aes(label = c(paste0("309 stops along 9 fixed routes", 
                                 "\n\n29 temporally-balanced samples", 
                                 "\ncollected over the course of 1 year"  
                                  )),
            x=-122.155, #x=-122.08, 
            y=47.714,  
            hjust=0, 
            ), size=text_size, 
            col = "#4b2e83"
            ) +
  geom_label(aes(
    #label = c("UFP\nBC\nNO2\nPM2.5\nCO2"), 
    #x=-121.935, y=47.621, 
     label = c("UFP", "BC", "NO[2]", "PM[2.5]", "CO[2]"), 
    x=-121.99, y=c(seq(47.665, 47.567,  length.out = 5)),
    hjust=0,), size=text_size, col = "#4b2e83", parse=TRUE
    ) + 
  #only keep map outline
  theme_void()  
  #ylab(expression(Anthropogenic~SO[4]^{"2-"}~(ngm^-3)))
  #theme_bw() #+
  #theme_linedraw()# +
  #theme_update(panel.background=element_blank())
  # theme_update(panel.background=element_blank(),
  # 
  #              panel.grid.major = element_blank(),
  #              panel.grid.minor = element_blank(),
  # 
  #              #panel.border = element_rect(colour = "#4b2e83", fill=NA,)
  #              )
  

ggsave(file.path("..", "Manuscript", "Images", "toc_map.jpg"), width = 3.25, height = 1.75,dpi = 300, units = "in")


```

 



# Density Plots


```{r}
# covars <- c("log_m_to_a1", #"log_m_to_a123", 
#             "log_m_to_l_airp", #"log_m_to_airp", 
#             "log_m_to_rr", "pop10_s01000", "elev_elevation")

covars <- c("log_m_to_a1", "log_m_to_a1_a2_intersect",  
            "log_m_to_l_airp",
            "log_m_to_ry", 
            "log_m_to_l_port",
            "log_m_to_coast"
            )

saveRDS(covars, file.path("Data", "Output", "density_covars.rda"))

mm_act <- mm_shp %>%
  select(covars) %>%
  mutate(Kind = "Mobile Monitoring Stops") %>%
  rbind(mutate(select(act_shp, covars), Kind = "ACT Cohort")) %>%
  st_drop_geometry()


mm_act %>%
  gather("Covariate", "Value", -Kind) %>%
  mutate(
    Covariate = recode_factor(factor(Covariate),
                              "log_m_to_a1" = "Log Meters to A1 Road", 
                              "log_m_to_a1_a2_intersect" = "Log Meters to A1-A2 Intersection",
                              "log_m_to_coast" = "Log Meters to Coast",
                              "log_m_to_l_airp" = "Log Meters to Large Airport (Sea-Tac Int.)",
                              
                              "log_m_to_l_port" = "Log Meters to Large Port (Port of Seattle)",
                              "log_m_to_ry" = "Log Meters to Railyard"
                              ),
    Covariate = gsub("Log Meters to ", "", Covariate),
    Value = exp(Value)
    
  ) %>%
  
  ggplot(aes(x=Value, fill=Kind)) + 
  facet_wrap(~Covariate) + 
  geom_density(alpha=0.3) + 
  theme(legend.position = "bottom",
        strip.text.x = element_text(size = 15)
        ) +
  labs(fill = "Location",
       x = "Proximity (Meters)"
       )
   
ggsave(file.path("..", "Manuscript", "Images", "density_plots.jpg"), width = 10, height = 8)

```





# Session Info

```{r}
#R.Version()$version.string

si <- sessionInfo() 

sessionInfo()   

# si[1]
# 
# si2 <- si#$otherPkgs

si$running


#names(si2) 
#test3 <- si2[1]  

print(si, RNG = TRUE, locale = FALSE)
#toLatex(si, locale = FALSE) # shortest

```
